requirements.txt
kafka-python
flask